Implement structureStack and improve parser test coverage
The structureStack now records a history of elements
that were added to and popped from it. This history
is used in testing, to aid developers when a parser
function failed to complete the structureStack.

The small test framework (for parser functions) now
has an improved error messages for non-complete
structureStacks.


Former-commit-id: 8f31f9107d5ea8475a3645ee040d41592fd44c25