Extract structureStack and improve error messages
The structureStack now records a history of elements
that were added to and popped from it. This history
is used in testing, to aid developers when a parser
function failed to complete the structureStack.

The small test framework (for parser functions) now
has an improved error messages for non-complete
structureStacks.


Former-commit-id: 0a38d20ed7d4899837da1171a251f826c039135f